requirements.txt
anthropic
requests
python-dotenv
ddgs
